compiled, seperated script from util to diagnostic.py
