+++ met-scDRS-method/src/met_scdrs/__init__.py
@@ -1,6 +1,6 @@
-from . import method, util, version
+from . import method, util, version, diagnostic
 from .method import score_cell
 from .preprocess import normalize, preprocess
 from .version import __version__
 
-__all__ = ["method", "util", 'normalize', 'preprocess', 'score_cells']
+__all__ = ["method", "util", 'normalize', 'preprocess', 'diagnostic', 'score_cells']
